refactor(webgl): per-concern split of texture-management.ts
Third slice in the long-running @gjsify/webgl per-concern splits, sibling
to PR #309 (webgl2-rendering-context.ts 1245 -> 866 + 4 modules) and
PR #273 (the prior WebGL2 lifecycle slice). Same install*Methods(proto)
pattern, applied this time to the largest remaining base-class
method-group module under context/.

| Group                | New file                                       | LoC |
|----------------------|------------------------------------------------|----:|
| Lifecycle            | context/texture-management/lifecycle.ts        | 229 |
| 2D texture upload    | context/texture-management/tex-image.ts        | 335 |
| Copy + compressed    | context/texture-management/copy-and-compressed.ts | 184 |
| Parameters           | context/texture-management/parameters.ts       | 111 |
| Introspection        | context/texture-management/introspection.ts    | 187 |

Lifecycle: createTexture, deleteTexture, activeTexture, bindTexture,
pixelStorei + the _detachTextureFromAllFramebuffers helper.
2D texture upload: texImage2D, texSubImage2D including the
6-arg-vs-9-arg / 7-arg-vs-9-arg overload dispatch, the GdkPixbuf.Pixbuf
+ TexImageSource source extraction, the UNPACK_PREMULTIPLY_ALPHA_WEBGL
premultiplication step and the UNPACK_FLIP_Y_WEBGL row-reverse step.
Copy + compressed: copyTexImage2D, copyTexSubImage2D,
compressedTexImage2D, compressedTexSubImage2D.
Parameters: texParameterf, texParameteri, getTexParameter +
_getTexParameterDirect (bypasses the parameter-validation switch).
Introspection: every _-prefixed helper the other modules reach for
(_getActiveTextureUnit, _getActiveTexture, _getTexImage,
_checkTextureTarget, _validTextureTarget, _validCubeTarget,
_verifyTextureCompleteness, _computePixelSize, _computeRowStride,
_checkDimensions).

28 methods totalling ~880 LoC moved out of the base file verbatim.
Each new module declaration-merges its <Group>Methods interface into
WebGLContextBase via `declare module '../../webgl-context-base.js'` and
exports an install<Group>Methods(proto) function.

Base file context/texture-management.ts: 941 -> 61 LoC. It is now a
barrel + composite installTextureManagementMethods that fans out to all
five installs, plus a re-exported aggregate TextureManagementMethods
interface (extends all five) so existing imports of just
TextureManagementMethods from context/index.ts keep compiling. Each
per-concern install is imported twice -- once as a named import for the
call site and once as a bare side-effect import -- so tsc preserves the
augmentation in the emitted .d.ts (same technique as PR #309 / PR #273
/ the canvas2d-core split in PR #262).

context/index.ts's installAllContextMethods(WebGLContextBase.prototype)
aggregator is unchanged -- it still imports the single
installTextureManagementMethods from texture-management.ts, which now
fans out internally.

The non-trivial cross-module call graph is preserved verbatim:
texImage2D / texSubImage2D reach into introspection (_getTexImage,
_computePixelSize, _computeRowStride, _checkDimensions) and call
_updateFramebufferAttachments from the base class; texParameter{f,i} /
getTexParameter reach into introspection (_checkTextureTarget,
_verifyTextureCompleteness); _verifyTextureCompleteness calls back into
the lifecycle module's bindTexture; deleteTexture calls back into the
lifecycle module's _detachTextureFromAllFramebuffers.

tsc --noEmit clean across @gjsify/webgl. 860 GJS webgl conformance
tests still green (no count change).
